Move /pkg/api to /apis
