+++ apis/doc.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package api
+package apis
 
 import (
 	cmapi "github.com/jetstack/cert-manager/pkg/api"
@@ -22,7 +22,7 @@ import (
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 
-	policycertmanageriov1alpha1 "github.com/cert-manager/policy-approver/pkg/api/v1alpha1"
+	pcmapi "github.com/cert-manager/policy-approver/apis/v1alpha1"
 )
 
 var (
@@ -31,6 +31,6 @@ var (
 
 func init() {
 	utilruntime.Must(clientgoscheme.AddToScheme(Scheme))
-	utilruntime.Must(policycertmanageriov1alpha1.AddToScheme(Scheme))
+	utilruntime.Must(pcmapi.AddToScheme(Scheme))
 	utilruntime.Must(cmapi.AddToScheme(Scheme))
 }
